+++ src/main/java/chess/domain/piece/Empty.java
@@ -1,4 +1,4 @@
-package chess.domain.chessboard.state;
+package chess.domain.piece;
 
 import chess.domain.chessboard.Coordinate;
 import chess.domain.chessboard.Square;
