+++ src/main/java/chess/domain/piece/state/Bishop.java
@@ -1,7 +1,8 @@
-package chess.domain.chessboard.state.piece;
+package chess.domain.piece.state;
 
 import chess.domain.chessboard.Coordinate;
-import chess.domain.chessboard.state.Team;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Team;
 import java.util.List;
 
 public final class Bishop extends Piece {
