+++ src/main/java/chess/domain/piece/state/Queen.java
@@ -1,12 +1,13 @@
-package chess.domain.chessboard.state.piece;
+package chess.domain.piece.state;
 
 import chess.domain.chessboard.Coordinate;
-import chess.domain.chessboard.state.Team;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Team;
 import java.util.List;
 
-public final class Rook extends Piece {
+public final class Queen extends Piece {
 
-    public Rook(final Team team) {
+    public Queen(final Team team) {
         super(team);
     }
 
@@ -14,16 +15,24 @@ public final class Rook extends Piece {
     public List<Coordinate> findRoute(final Coordinate from, final Coordinate to) {
         validatePossibleDestination(from, to);
 
+        if (from.isPositiveDiagonal(to)) {
+            return positiveDiagonalRoute(from, to);
+        }
+
         if (from.isSameFile(to)) {
             return verticalRoute(from, to);
+        }
+
+        if (from.isNegativeDiagonal(to)) {
+            return negativeDiagonalRoute(from, to);
         }
 
         return horizontalRoute(from, to);
     }
 
-
     private void validatePossibleDestination(final Coordinate from, final Coordinate to) {
-        if (!(from.isSameFile(to) || from.isSameRank(to))) {
+        if (!(from.isSameFile(to) || from.isSameRank(to) || from.isPositiveDiagonal(to) || from.isNegativeDiagonal(
+                to))) {
             throwCanNotMoveException();
         }
     }
