+++ src/main/java/chess/domain/piece/state/Rook.java
@@ -1,12 +1,13 @@
-package chess.domain.chessboard.state.piece;
+package chess.domain.piece.state;
 
 import chess.domain.chessboard.Coordinate;
-import chess.domain.chessboard.state.Team;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Team;
 import java.util.List;
 
-public final class Queen extends Piece {
+public final class Rook extends Piece {
 
-    public Queen(final Team team) {
+    public Rook(final Team team) {
         super(team);
     }
 
@@ -14,23 +15,16 @@ public final class Queen extends Piece {
     public List<Coordinate> findRoute(final Coordinate from, final Coordinate to) {
         validatePossibleDestination(from, to);
 
-        if (from.isPositiveDiagonal(to)) {
-            return positiveDiagonalRoute(from, to);
-        }
-
         if (from.isSameFile(to)) {
             return verticalRoute(from, to);
-        }
-
-        if (from.isNegativeDiagonal(to)) {
-            return negativeDiagonalRoute(from, to);
         }
 
         return horizontalRoute(from, to);
     }
 
+
     private void validatePossibleDestination(final Coordinate from, final Coordinate to) {
-        if (!(from.isSameFile(to) || from.isSameRank(to) || from.isPositiveDiagonal(to) || from.isNegativeDiagonal(to))) {
+        if (!(from.isSameFile(to) || from.isSameRank(to))) {
             throwCanNotMoveException();
         }
     }
